+++ tools/daemon/hup.do.pl
@@ -8,7 +8,7 @@
 # @(-) --name=<name>           the daemon name [${name}]
 # @(-) --port=<port>           the port number to address [${port}]
 #
-# The Tools Project - Tools System and Working Paradigm for IT Production
+# TheToolsProject - Tools System and Working Paradigm for IT Production
 # Copyright (©) 1998-2023 Pierre Wieser (see AUTHORS)
 # Copyright (©) 2023-2025 PWI Consulting
 #