+++ tools/dbms/sql.do.pl
@@ -19,7 +19,7 @@
 # @(@) Use Ctrl+D in a Unix terminal.
 # @(@) '--dummy' option is ignored when SQL command is a SELECT sentence.
 #
-# The Tools Project - Tools System and Working Paradigm for IT Production
+# TheToolsProject - Tools System and Working Paradigm for IT Production
 # Copyright (©) 1998-2023 Pierre Wieser (see AUTHORS)
 # Copyright (©) 2023-2025 PWI Consulting
 #