+++ tools/ttp/push.do.pl
@@ -12,7 +12,7 @@
 #
 # @(@) When specified, the '--exclude-dir' and '--exclude-file' options override the corresponding values from the site configuration.
 #
-# The Tools Project - Tools System and Working Paradigm for IT Production
+# TheToolsProject - Tools System and Working Paradigm for IT Production
 # Copyright (©) 1998-2023 Pierre Wieser (see AUTHORS)
 # Copyright (©) 2023-2025 PWI Consulting
 #