+++ ChangeLog.md
@@ -9,6 +9,7 @@
     - TTP::Metric enabled default value is true
     - TTP::Metric replace 'Telemetry' obsoleted value with 'telemetry'
     - Define new TTP::Telemetry::isHttpEnabled() isMqttEnabled() and isTextEnabled() functions
+    - TTP::Metric has labelled error codes
 
 ### 4.9.1