+++ ChangeLog.md
@@ -11,6 +11,7 @@
     - Define new TTP::Telemetry::isHttpEnabled() isMqttEnabled() and isTextEnabled() functions
     - TTP::Metric has labelled error codes
     - Make project name and copyright notice consistent among all files (todo #133, #134)
+    - Remove unused ttp.pl test verb (todo #131)
 
 ### 4.9.1