+++ ChangeLog.md
@@ -1,4 +1,4 @@
-# The Tools Project - Tools System and Working Paradigm for IT Production
+# TheToolsProject - Tools System and Working Paradigm for IT Production
 
 ## ChangeLog
 
@@ -10,6 +10,8 @@
     - TTP::Metric replace 'Telemetry' obsoleted value with 'telemetry'
     - Define new TTP::Telemetry::isHttpEnabled() isMqttEnabled() and isTextEnabled() functions
     - TTP::Metric has labelled error codes
+    - Make project name and copyright notice consistent among all files (todo #133, #134)
+    - Remove unused ttp.pl test verb (todo #131)
 
 ### 4.9.1